test 16

+++ controllers/cards.js
@@ -16,32 +16,30 @@ const getCards = (req, res) => Card.find()
     message: 'Ошибка сервера',
   }));
 
-const deleteCardById = (req, res) => {
+const deleteCardById = (req, res, next) => {
   const { id } = req.params;
   if (!isValidId(id)) {
     return res.status(400).send({
       message: 'Передан некорректный ID',
     });
   }
-  return Card.findByIdAndDelete(id)
+  return Card.findById(id)
     .then((card) => {
       if (!card) {
         return res.status(404).send({
           message: 'Карточка не найдена',
         });
       }
-      return res.status(200).send(card);
-    })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        return res.status(400).send({
-          message: 'Карточка не найдена',
-        });
+      const cardOwner = card.owner.toString();
+      if (cardOwner === req.user._id) {
+        return Card.findByIdAndDelete(id)
+          .then(() => res.status(200).send(card));
       }
-      return res.status(500).send({
-        message: 'Ошибка сервера',
+      return res.status(403).send({
+        message: 'Вы не автор карточки',
       });
-    });
+    })
+    .catch(next);
 };
 
 const createCard = (req, res) => Card.create({ ...req.body }).then((card) => {
